Fix `Timestamp.to_msec`.
I've done my best to preserve the spirit of the old code, which appeared
to be trying to solve the whole "are we in era 0 or 1?" problem.

I'm guessing the use of `Bitwise.band(seconds, 16)` was meant to test
the most significant bits of the NTP time, but was accidentally testing
the least significant ones.  That meant the alternate epoch was
triggering roughly 50% of the time, when it shouldn't be triggering at
all until 2036.

Also, the era 1 epoch was wrong — it should have been negative, such
that subtracting it from the NTP time actually made the resulting Unix
time larger.

The result was that, a lot of the time, `to_msec` would give a time that
was off by ~123 million seconds (~3.8 years) in the future.  This would
result in a "Wrong originate timestamp" error, since the timestamps no
longer matched.  (In many cases, you could take the timestamp we send in
our request packet in `SNTP.Socket`, run it through `to_msec`, and get
the completely wrong time.)

This dramatically simplifies the existing code, and should work
correctly until the early 2100s, by which time hopefully we'll have
64-bit NTP anyway.